adjust - separa tests em arquivos por classificação
